requirements.txt
django==4.2.11
djangorestframework==3.15.0
psycopg2-binary==2.9.9
pillow==10.2.0
django-filter
django-sqlprint-middleware
django-extensions
ipython
drf-yasg
drf-spectacular
django-mptt